fix bugs with marker work

+++ laravel/resources/js/components/request-services/request-service.js
@@ -8,15 +8,15 @@ export class RequestService {
 
 
     update(url, _body) {
-        return axios.patch(url, _body);
+        if (url && _body) return axios.patch(url, _body);
     }
 
     destroy(url) {
-       return axios.delete(url);
+        if (url) return axios.delete(url);
     }
 
     store(url, _body) {
-        return axios.post(url, _body)
+        if (url && _body) return axios.post(url, _body)
     }
 
 }
